Debugging and refactoring
