+++ Makefile
@@ -12,7 +12,7 @@ HTSLIB    = $(HTSLIBDIR)/libhts.a
 HTSLIBINCLUDE = $(HTSLIBDIR)
 HTSLIB_VERSION = b6aa0e6
 
-ofiles    = src/main.cpp.o src/misc.cpp.o src/params.cpp.o
+ofiles    = src/main.cpp.o src/misc.cpp.o src/params.cpp.o src/io.cpp.o
 hfiles    = $(wildcard src/*.h)
 
 .PHONY: all clean cleanall install readme format test
