+++ src/misc.cpp
@@ -35,49 +35,6 @@ void ococo::print_version() {
     std::cerr << std::endl;
 }
 
-void ococo::fatal_error(const char *format, ...) {
-    va_list args;
-    va_start(args, format);
-    fprintf(stderr, "[ococo:fatal-error]: ");
-    vfprintf(stderr, format, args);
-    va_end(args);
-}
-
-void ococo::error(const char *format, ...) {
-    va_list args;
-    va_start(args, format);
-    fprintf(stderr, "[ococo:error]: ");
-    vfprintf(stderr, format, args);
-    va_end(args);
-}
-
-void ococo::warning(const char *format, ...) {
-    va_list args;
-    va_start(args, format);
-    fprintf(stderr, "[ococo:warning]: ");
-    vfprintf(stderr, format, args);
-    va_end(args);
-}
-
-void ococo::info(const char *format, ...) {
-    va_list args;
-    va_start(args, format);
-    fprintf(stderr, "[ococo]: ");
-    vfprintf(stderr, format, args);
-    va_end(args);
-}
-
-bool ococo::file_exists(const std::string &fn) {
-    FILE *file;
-
-    file = fopen(fn.c_str(), "r");
-    if (file) {
-        fclose(file);
-        return true;
-    }
-    return false;
-}
-
 double ococo::realtime() {
     struct timeval tp;
     // struct timezone tzp;
