+++ src/stats.h
@@ -393,7 +393,7 @@ int stats_t<T>::call_consensus_position_uncompressed(FILE *vcf_file,
 
     if (vcf_file != nullptr) {
         if (old_base_nt256 != new_base_nt256 || params->verbose) {
-            print_vcf_substitution(vcf_file, seq_name[seqid], pos,
+            ococo::print_vcf_substitution(vcf_file, seq_name[seqid], pos,
                                    old_base_nt256, new_base_nt256, psu);
         }
     }
